:sparkles: parsing done, with norm check
*
