+++ src/lexer/copy.c
@@ -12,45 +12,6 @@
 
 #include "minishell.h"
 
-int	is_delimiter(char *str)
-{
-	char	*special;
-	char	*operand;
-	int		i;
-
-	i = 0;
-	special = "|<>\'\"()$";
-	operand = "&&";
-	while (str && str[i] && (str[i] > 32 && str[i] < 127) \
-		&& !strchr(special, str[i]) && ft_strncmp(operand, &str[i], 2))
-		i++;
-	return (i);
-}
-
-int	is_delimiter_spc(char *str)
-{
-	char	*special;
-	char	*operand;
-	int		i;
-
-	i = 0;
-	special = "|<>\'\"()";
-	operand = "&&";
-	while (str && str[i] && (str[i] > 31 && str[i] < 127) \
-		&& !strchr(special, str[i]) && ft_strncmp(operand, &str[i], 2))
-	{
-		if (str[i] == '$' && is_delimiter_spc(&str[i + 1]))
-			break ;
-		else if (str[i] == ' ' && str[i + 1] > 32 && str[i + 1] < 127)
-		{
-			i++;
-			break ;
-		}
-		i++;
-	}
-	return (i);
-}
-
 char	*copy_chars(char *str, int *i, int n)
 {
 	char	*value;
