+++ src/lexer/delimiter.c
@@ -1,0 +1,52 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   delimiter.c                                        :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: foctavia <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/09/08 11:34:44 by foctavia          #+#    #+#             */
+/*   Updated: 2022/09/08 11:34:46 by foctavia         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "minishell.h"
+
+int	is_delimiter(char *str)
+{
+	char	*special;
+	char	*operand;
+	int		i;
+
+	i = 0;
+	special = "|<>\'\"()$";
+	operand = "&&";
+	while (str && str[i] && (str[i] > 32 && str[i] < 127) \
+		&& !strchr(special, str[i]) && ft_strncmp(operand, &str[i], 2))
+		i++;
+	return (i);
+}
+
+int	is_delimiter_spc(char *str)
+{
+	char	*special;
+	char	*operand;
+	int		i;
+
+	i = 0;
+	special = "|<>\'\"()";
+	operand = "&&";
+	while (str && str[i] && (str[i] > 31 && str[i] < 127) \
+		&& !strchr(special, str[i]) && ft_strncmp(operand, &str[i], 2))
+	{
+		if (str[i] == '$' && is_delimiter_spc(&str[i + 1]))
+			break ;
+		else if (str[i] == ' ' && str[i + 1] > 32 && str[i + 1] < 127)
+		{
+			i++;
+			break ;
+		}
+		i++;
+	}
+	return (i);
+}
